+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 1.0.0-alpha.7
+- Fixed lobby refresh not triggered after entering a lobby after creating one
+- Added back button on lobby menu
+
+# 1.0.0-alpha.6
+- Experimental create game on network menu changes
+
+# 1.0.0-alpha.5
+- Fixed lobby update issue message at changing between menus
+- Fixed issue with input autofocus when opening menus
+
+# 1.0.0-alpha.4
+- Minor fixes
+- Design changes
+
 # 1.0.0-alpha.3
 - Fixed list navigation on multiplayer menu options
 - Added chimera history and bookmark servers screen